SQL Agent 2/requirements.txt
fastapi
uvicorn
langchain
ollama
python-dotenv
requests
aiohttp
